Improve auth form handling of JS errors
By using POST instead of GET we prevent critical data from appearing
in the URL when a JS error occurs (or JS is disabled).

Using the jQuery preventDefault() method is better than 'return false'
because it does not stop propagation of the event, plus it can be
called at the beginning of the event handler and therefore prevent
submission of the form on JS error.

Note: The server is unable to handle the forms without JS, otherwise
it would make more sense to place preventDefault() at the end of the
event handler.

closes #12492

Change-Id: Ib62ba1defa50ebcff94a3ef2487ec142838620f6