Fix linotpAudit during LSE Smart Virtual Appliance installation
Since linotpAudit.sql.url was commented out the postinst script didn't
replace the line correctly and and error was raised when running 'paster
setup-app'.

Now linotpAudit.sql.url is always replaced (with or without comment) when
the installer detects that it is running on a SVA. On other
(Debian/Ubuntu) systems the Audit is not enabled by default since it
reduces performance and not everybody wants to use it. It can be manually
enabled after installation.

We should consider asking the user if he/she wants to enable the Audit Log
during installation.

Change-Id: Id38cdf428807ac01b8096d49e89b26183ed607c7