During installation move Apache2 config files to etc/linotp2
Until now these files were directly moved to etc/apache2/sites-available
(where the corresponding etc directory is /etc on Debian or a plain
directory where the LinOTP package is installed on other systems).

The reason for not putting them directly in the sites-available is to
support different versions of Apache2. The config format changes
significantly between Apache2.2 and Apache2.4 .

Now all Apache2 config files are moved to etc/linotp2/apache2.2-example
and etc/linotp2/apache2.4-example. The Debian postinst script then moves
the corresponding VirtualHost file to sites-available and enables it.

The Apache 2.4 config is just a copy of the Apache 2.2 config, replacing
the 'Order' directives with the new 'Require' directive.

refs #12518
refs #12521

Change-Id: I188c0e07e05fd207e9fd2f73846680cffb0fa8a2