Fixed getPolicy when specifying an exact action
The previous behaviour was to match the first part of the action (before
the =) to the parameter passed into the function.

For example:
Assuming we have a policy stored in the DB with action: otppin=1

Calling getPolicy with param action 'otppin=1' would return no results.
Calling getPolicy with param action 'otppin' would return the correct
result.

This behaviour came to light because the controller system/getPolicy
used to ignore the 'action' and 'user' parameters. But since commit
27e6ca3887953fa6dddd0c25ca005c7acfef95cb this is no longer the case.

This means that when searching for a policy with exactly the same
parameters used to create it (as done by test_authorize tests) following
happens:
- Before commit 27e6ca388 the correct result was returned because
  'action' was ignored.
- After commit 27e6ca388 no results were returned because the action was
  explicit (otppin=1) and therefore didn't match. Passing in just 'otppin'
  would have worked.
- After the current commit explicit actions are also matched and the
  correct result is returned

The behaviour implemented by this commit is as follows:
Assuming 3 policies are stored in the DB with following actions:
Policy1: otppin=1
Policy2: otppin=2
Policy3: otppin=0

Calling getPolicy({'action': 'otppin=1'}) will return Policy1
Calling getPolicy({'action': 'otppin'}) will return all three policies.

Change-Id: Iaa32641846cf01b8938e3f824cb2cb948908bf52