reworked support for user as parametere to system/getPolicy
as the policy/getPolicy has a bug with empty user in definition
we make the search wo. user and postprocess the result, as there
is no resctriction derived from the result. The verification
is done against the get_client_policy, which has the same behaviour.

The reason for this support has come from the requirement to inform
the UI which policy for random otp pin is matching and is implemented
against the getRandomOTPPINLength which is used in the server evaluation
of the policy (which usese the get_client_policy).

Change-Id: Ia0394897b88cd512d9179fd7d433fbbf7b462588