fix the breaking of token enrollment due to fail in create qrcode
the last code change showed that the qrcode create in the tokenclass.py
was wrong handled - the googleurl++ should only be part of the
hmac and totp token class. As well the qrcode gen in the ocra tokens
was incomplet as during the first step in enrollment the otpkey is
not set and thus empty and thus must not be displayed.

Change-Id: I58a45c8932671fbeba31f27dfaeceed4be5fbfe8