Move block that fixes ask_createdb issue
On Ubuntu 12.04 the debian/config check_mysql_tables() function has
different results causing a different path in the code to be taken which
means that the debconf FSET is not executed.

Now FSET is always executed when using a MySQL database, and upgrading
from a LinOTP version between 2.6 (inclusive) and 2.7.1 .

Also fix a small issue that could have lead to big problems: dpkg
--compare-versions returns a non-zero value when the check fails. Since
errexit is set this means that the whole config script would die,
preventing the package from being installed. By surrounding the block
with 'set +e' and 'set -e' this issue is prevented.

closes #12718

Change-Id: Id766a51473ec830ac6a7b93ea4e19bf45def8b6f