Add config option 'pkcs11.accept_invalid_padding'
The pad() method in the pkcs11 module published with LinOTP version <
2.7.1 padded data incorrectly when its length was a multiple of the
block length.

For this reason we are introducing this config option that will allow
incorrectly padded data to be used.

If you are upgrading from LinOTP < 2.7.1 AND use a PKCS #11 HSM it
probably is a good idea to set this option to True in the LinOTP config.
Otherwise simply ignore it.

A similar config option was already implemented in commit 4669ac9a04c.
We decided to remove it in commit 6dff0bfba6a and are now reverting this
decision because we need the option to migrate old systems.

refs #12575

Change-Id: Iddb69538b6676cbcb6143cbcb32487f0b1c82c0f