Fix for the resolving of unknown users
In case of an unknown user the realm is marked as wildcard
in a first place.

The problem is that resolving of the users stops here. If
the resolver does resolve to no user, all tokens to this
'none' user are returned, which shows as all unassigned
tokens.

The fix iterates over all resolvers to identify all potential
users and raises an exception if no user could be found.

Change-Id: I11b4e26035723a085f48fb3faba128de7fd2b218