Copy correct Apache2 VirtualHost in LinOTP Debian postinst
Depending on the Apache2 Version installed on the machine the format of
the VirtualHost config file changes. We only check for Apache >= 2.4
because the format change was introduced with Apache 2.4 . If future
releases of Apache2 introduce new changes the script needs to be
adapted.

In the postinst script the installed Version of Apache2 is checked and
either one or another file from /etc/linotp2 is copied to /etc/apache2 .

The example VirtualHost files copied by the script are deposited under
/etc/linotp2 by the setup.py script during installation (done
transparently at the beginning of the Debian package installation).

refs #12518
refs #12521

Change-Id: I98d68027de64953de1dfc1f31316cebf8daba8b4