fix failing unicocde test in validate
due to the fact, that we now try to set the base_auth_user in the
base class of the controllers, here as well unicode decode errors
might happen. Best way to not break compatibility is to ignore
the exception as this will be handled in the contollers anyway
but there the corresponding response could be composed.

Change-Id: I6030b2e4ec9922de6231f384b1500f8200951bcf