linotp-auth-radius: Encode user and password as UTF-8 before sending
Commit 7ac9b48 made sure that 'user' and 'password' are converted to
Unicode for internal handling (after reading from the command-line).

This works fine with pyrad 2.0 because data is packed with struct.pack
before sending it over the network.  struct.pack correctly handles
Unicode strings.

pyrad 1.2 on the other hand expects a regular string (bytes) and will
therefore not handle Unicode strings properly.

By explicitly encoding the data as UTF-8 before passing it on to pyrad
we make sure it works in both pyrad versions.

Change-Id: I8ddf22836de3c08af9394398d2873180882997d5