added ldap proxy support - to support the access of an AD through slapd
if object type is guid (only AD) there is a special query to AD to
direct access these guid objects via base='<guid=UserId>'. But as this
is not part of standard LDAP, this access will fail if access is routed
through a proxy. The fix allows to specify in the LDAP configuration
that a proxy is used in between by a boolean switch.

Change-Id: Ic7480cef67a123be678af6f143c1a0e769428abb