Remove all linotp. entries from test.ini
Since the config is stored in the DB and sometimes read into memory you
cannot trust values that are not in memory. The DB is destroyed after
every test module, but test.ini is only read ONCE at the beginning.

The config in memory should be overwritten before every test module (at
least in the parts where linotp is concerned).

Change-Id: I4cca72e75f520de80b2d704d558fc84112da2d16