Fix 'httperror' handling in sendError function
In WebOb versions >= 1.2 if some request parameter can not be decoded
correctly an UnicodeDecodeError is raised. Before WebOb 1.2 this was not
an issue because the second parameter 'errors' of the 'decode' function
was set to 'replace', causing incorrect values to be replaced by the the
official U+FFFD REPLACEMENT CHARACTER.

Currently this leads to the 'httperror' variable being assigned a string
representation of that exception. Since 'httperror' is later used to set
the HTTP Status Code this leads to unexpected results because the Status
Code is not numeric.

This commit tries to fix this issue by:
1. Trying to 'manually' extract the 'httperror' variable from the
   parameters if the default Pylons/WebOb method fails.
2. Always setting 'httperror' to a valid value:
     - None
     - The string representation of an integer

The issue of checking if 'httperror' is actually a valid HTTP Status
Code (i.e. 1xx - 5xx) was not addressed before and is NOT part of this
commit.

refs #12804

Change-Id: I437deb9da1a4f55fe3f40304b01b0621a3696c65