Don't deactivate linotp2 VirtualHost during upgrade (Apache 2.4)
Currently the postinst script activates the linotp2 Apache VirtualHost
during installation or upgrade. To make sure it is not activated when it
was manually disabled the prerm script (executed before postinst during
upgrade) checks if the linotp2 file exists under sites-enabled. Since
with Apache 2.4 the file was renamed to linotp2.conf the check
incorrectly assumes the site was manually disabled and does not enable
it in the postinst script.

closes #12806

Change-Id: I0eb84c6a6c208869d25919cbcd0263875c504d54