Support older WebTest versions in functional tests
On Debian Squeeze the python-webtest package has version 1.2.1

In test_manage.py the 'testbody' property of WebTest Response is used.
This property was introduced in version 1.2.2

Since the Response class defines a __contains__ method it is actually
easier to just assert 'abc' in response. Whitespace is normalized, which
simplifies the test.

Change-Id: I1a919f8129cf8f618f3d980df07a1669f5a3937a