Add support for valid_facets in the U2FTokenClass
A new u2f controller is added with an action called 'valid_facets'. This
controller returns a JSON list containing all valid facets specified by the
enrollment policy 'u2f_valid_facets'.  The URL to this action can be sent to
the U2F client (eg. the Chrome extension) to allow usage of the registered U2F
token for all listed facets.

The handling of the U2FTokenClass is updated to work with these valid facets as
well as without them. In the second case the origin parameter extracted from
the client data object is saved to the TokenInfo, blocking other values of
origin in later authentications.

Change-Id: If9b5e66bd5a7f7c9bb4b66e55613fbd83a15e77b