Specify RADIUS auth and acct ports in tests to enable parallelization
In order to run several tests at the same time on the same machine the
dummy_radius_server tool needs to be started on different ports for
every test.

The ports to be used are passed in using the nose-testconfig plugin.

If nose-testconfig is not installed or the necessary values are not set
then default values are used to ease running the tests.

Change-Id: I6943e0347c4988a293ce2121717b3d5953b23a92