Specify 'paster' port in tests to enable parallelization
In order to run several tests at the same time on the same machine the
'paster' LinOTP server needs to be started on different ports for every
test.

The ports to be used are passed in using the nose-testconfig plugin.

If nose-testconfig is not installed or the necessary values are not set
then default values are used to ease running the tests.

Change-Id: Ibbb8a1f4c0e2a7a8f23a0c3884fbc4b62376fdd2