support for autoassign with remote token and yubikey
in case of the remote token only the remote token would be
autoassigned. But as the real passwd is required for the
assignement, the target token as well could be autoassigned.

For the remote token to support autoassignment the method
check_otp_exist was added.

To support the autoassignment with remote tokens (s.o.)  one
additional enrollment policy was defined: autoassigne_forward
(bool), which controlles, if the target token will recieve the
same password and user but remote.realm.

In addition one bug token search bug was fixed, which controlls,
that the token to be found must be in the same realm as the user
not only the resolver must match for the token!!

Addition to the web ui of the remote token was made, so it is
possible to define the otppin length which is required if the
token pin check is done localy (and remote does no pin check or
has a fixed pin+pass length)

Conflicts:
	.gitreview
	linotpd/src/linotp/lib/token.py

Change-Id: Ie812c52265e92b6fedf5c24fabf1f67710c17746