Add SQLAlchemy<=0.9.99 dependency due to 'SQLAlchemy Migrate'
The latest 'SQLAlchemy Migrate' version (0.9.6) is not compatible with
SQLAlchemy version 1.0.0
This dependency is declared in 'SQLAlchemy Migrate' but it is not
honoured during installation, leading to errors when trying to start the
LinOTP server.

Until this problem is fixed in 'SQLAlchemy Migrate' (or pbr or
setuptools?) we pin the SQLAlchemy version.

refs #12848

Change-Id: I8d3e6db2c50f5ab0ea495a5e987e97c2a7862023