Fix DOING_* helper variables in debian postinst and config
Because of the way the Debian postinst script is called it is actually
not possible to distinguish between an upgrade and a new installation
the way it was done in this script until now.

Uninstalling (apt-get remove) and then re-installing the package looks
the same as a package up- or downgrade. This is because the second
parameter passed to the script is the last version of the package that
was configured and this information is kept when simply removing
(apt-get remove) the package. In case of a purge (apt-get purge) this
information is removed and a new installation actually looks like a new
installation.

The variables were renamed and adapted to reflect this new understanding
of Debian installer scripts.

The config script was similarly adapted and some comments were added on
how and when the script is called.

In both scripts variables to track the current version being installed
were added. These variables need to be adapted for every release. The
(internal) BuildPreprocessor tool does this automatically
(build_preprocess set-version).

Concerning functionality nothing should have changed.

closes #12872

Change-Id: I7cb64402f817923faf87029ca239bebd18c1d8cc