Debian jessie: Allow webob version 1.4
In debian jessie, pylons 1.0.1 is shipped together with webob 1.4. LinOTP
requires webob < 1.4 due to incompatibilities between pylons and webob.
However, jessie ships with a patched version of pylons working with webob in
version 1.4.
This commit includes a quilt patch in the jessie subfolder in the
debian/patches directory, which removes the webob < 1.4 dependency.
It should only be applied during a build on jessie.

closes #12951

Change-Id: I82cc921d463a1138434cd44943e3824587dd288f