Fix delivering correct language
In Internet Explorer 11, the default german language code is 'de-DE' and not
'de'. Pylons' set_lang function of the i18n.translation module needs a
locale name as parameter and therefore can't handle the dash in 'de-DE'.

The HTTP-ACCEPT-LANGUAGE header is now parsed using a regular expression and
the requested language code is transformed to a locale name before being sent
to the set_lang function.

closes #12859

Conflicts:
	linotpd/src/linotp/lib/base.py

Change-Id: I5d9172d4b0066ccd95d424a74bb48204a8f4612b