Support for multiple FIDO U2F tokens per user
If a challenge is requested and the user owns multiple U2F tokens, one
challenge per token will be created and sent to the user. Once a response is
received it is checked against each of the tokens with a pending challenge.
Therefore we should return a proper return value to the calling function and
not raise exceptions on a challenge or signature mismatch during the response
check.

Tests for multiple tokens per user are added to the U2F functional tests.

closes #12960

Change-Id: Ia640fd915166dd199d2a8a86b9f3c9f67074f874