Remove trailing slash in /manage/ (Selenium Tests)
This caused some strange behaviour (such as all buttons getting
continously bigger) when running the tests on Mozilla Firefox, and
recently also on Chromium.

The exact cause of the problem is still unclear. It can be reproduced by
opening the WebUI with the Digest credentials directly in the URL (e.g.
https://admin:Test123!@FQDN/manage/ ).

closes #13014

Change-Id: I7e1dd4a2f9b8704c94b65c61e703676fc310416e