support for autoassignment without action value
the autoassignment in its first version defined an otplen
of the target token. with the 2.6/2.7 version this value
was not required anymore, as for every token in scope, the
otplen has been checked first before extacting the password.

By this patch the autoassignment is now a boolean policy and the
getActionValue now support these too. This guaranties that both
cases, with int value for otplen and boolean are supported.

closes #12879

Change-Id: Ic0e59647940165937d08d84b9983c5b082daf6d5