Changed Lost Token appearance in manage web ui; added email, sms
If you push the Lost token buttion in manage web ui a dialog shows up.
Before this commit three buttons to choose a lost token type was shown.
In order to improve the look and feel a drop down is implemented.
The drop down contains:
 - Password Token
 - Email Token
 - SMS Token
 - '- Select Temporary Token Type -' (is selected by default)
These fields are only selectable if the user which belongs to the token
has an email address and/or mobile (phone number) filled out.

Additionally a button named 'Get Temporary Token' is created to send up
the selected type.

closes #12894
closes #12814

Change-Id: Ib9e9b0ae406afe7d9320e1644594414e59a70728