Change rlm_linotp2 freeradius 'restrictusername' default to 'no'
If the 'restrictusername' option is set to 'yes' (as was the default
until now) then only usernames containing ASCII letters, numbers and a
few special characters -_+.@ were allowed.

In order to better support UTF-8 usernames and German special characters
(Umlaute) we change the default in the config file to 'no'.

In the code the default still is 'yes' so that existing systems won't
suddenly reverse their current behaviour after updating.

To restrict your usernames to the mentioned values you may either set
the value to 'yes' in the config file, or remove that line from the
config file.

refs #12862

Change-Id: Ic32807f0b3e0971e4a5119fb1f03d52acaa7ce25