Use Digest instead of Basic Auth in linotp-enroll-smstoken tool
Our linotp.deb package sets LinOTP up with Apache and Digest
authentication by default. This is also the setup recommended in our
documentation. Therefore it makes sense to use Digest by default.

In the future if required we could make this configurable by adding a
command-line option to choose different types of authentication.

closes #13048

Change-Id: I8bcc88e81f1e62636b89cb3090152fb9c19ef08c