Add test_policy2.py module with autoassignment functional tests
The reason for adding test_policy2.py instead of using the existing
test_policy.py is that test_policy.py is very messy and requires a very
specific test order. The new module can be kept cleaner from the
beginning and tests from the old module can slowly be cleaned up and
moved to the new one.

Change-Id: I452ca126e278f4792f13822d61eb8fbedab5901a