Don't delete non-existing policies in tests
In two cases policies were deleted that no longer exist.

In theory we could implement a delete_policy() method that does not fail
in this case (e.g. delete_policy_if_exists()) but for now I think it is
cleaner and gives testimony to a more defined behaviour if we avoid it.

Change-Id: I1a3e0307eaeebc2790dac2280dfb5a1ca8c6ad3e