Added new test-controller (called advanced test controller)
The new test-controller was designed as an improvement of the
existing one, with the idea to better support readability and
clients written in different programming languages (based on this new
controller we will implement portable test cases).

The new controller will hide the technical details about how to perform
LinOTP calls (pillons, authentication) and will provide methods for
direct LinOTP operations. The writer of test cases does not need to
understand how the communication works, he only need to know the LinOTP
interface-api and directly perform the operations which are needed in
tests.

More, the new controller will provide better support for checking the
return values and in all of the cases the result values are tested by
default (for success).

Change-Id: I84b936c181c2ba70c9ed13cff1a411b20e54569e