Improve advance TestController interface
1) proper implementation of $(get/set/delete)+Configuration.
Note: if value is None, the new setConfiguration method is performing a
delete (previously it was raising an error).

2) proper processing of expectedValues.
In order to implement more flexible code, the expectedValue is
preconfigured on a "DefaultValue" placeholder which, deppending on case,
will be initialized to a proper "value".

WARNING: because the 2 changes are related (the second change is used by
the first one) the 2 feature are commited together.

Change-Id: I82bfb8e1cc2c52f913b5a277a832ccf13db0843c