Simplify the U2F appId handling
To simplify the appId handling for the user, an enrollment policy 'u2f_app_id'
containing the appId can be defined. After defining the policy, no further
appId API parameters need to be passed to LinOTP.

This enrollment policy is optional and can be completely replaced by passing an
appId parameter to the API call creating a new token (phase registration1).
Following API calls for the created token (finishing up enrollment and
performing transactions) do not need an appId parameter.

If the appId policy is set and additionally an appId parameter is passed on
token creation, the API parameter will be used for creating the token.

closes #13272

Change-Id: Ie0a2094aba7b4b56adbc6f813d4f9fb3b098ef03