Append U2F appId to message object on triggered challenge
On triggering a challenge of a U2F token, a message object including the
challenge is sent to the user as response to the /validate/check API call. The
token's appId is appended to this message object to make it a U2F SignRequest
object, which can be parsed by the caller and passed to the U2F client without
further modification.

closes #13273

Change-Id: I2187f71e553bb56de72a388130c75ebdb18b1085