added test to check the qrcode to contain the lseqr url
the value entry in the ocraurl contains the lseqr://
url, which should be rendered into the qrcode. The test
assures that the value contains the lseqr:// url and
that the rendering of the lseqr:// url is the same as the
image data in the img entry under the ocraurl.

Conflicts:
	linotpd/src/linotp/tests/functional/test_ocra2.py

Change-Id: I7e0e187cde28d3dd9ded111c479cab8d210b30e5