SMSProvider: added some tests to check GET and POST requests for all httplibs
the httpsms provider supports several http libs, which are
requests, urllib and httplib. This test checks that for all
httplibs the GET and POST request work. in addition, the new
test for the HttpSMSProvider Twillo is implemented, which uses
the feature to check the smsprovider result against a regex expresion

Change-Id: I057a7e12afdebb47fef8d2972d5fa7ac2fcfbd39