SMSProvider: conversion of the phonnumber to become msisdn compliant
the phone numbers from the useridresolvers sometimes are not in the
format which is required on the sms provider side. The fallback of the
phone number format for those cases is the MSISDN format. To enforce the
the MSISDN format could be enforced in the sms provider config by the
parameter {"MSISDN":True}

Change-Id: I7e15b59d5a6ca3155dae69649e0e9d7405967afe