HTTSMSProvider conversion: urlencode expects utf8 instead of unicode
The urllib.urlencode expects utf8 for the http submission. In case that
the unicode parameters contains non ascii characters, the utf8
conversion fails. This fix applies the conversion from unicode to utf8
on all key and values of the parameters.
Remark: The method 'urlencode' has been ported from the branch-v2.7.2
and been extended with the utf8 conversion.
Todo: add unit tests for the url conversion in branch-v2.7.2

Conflicts:
	smsprovider/src/smsprovider/HttpSMSProvider.py

Change-Id: Ib5818ea14e63853087f6513aef4a66ccdcade6f1