freeradius: support for nas identifier instead of client ip
the nas identifiere is the ip behind the radius client. To support
the handling of the nas ip instead of the client ip similar to the
freeradius perl module, a configuration option has been added, which
is the boolean "prefer_nas_identifier", which is by default 'no' for
compatibility reasons.

tests have been made with the following cmd line:

  radtest -x 'brahms' 372403  radiusserver 0 geheim1 0 nas_originator

but only for the IPv4 code section!

Change-Id: Ibfcd70a09d7326fb71f3342329d3fbfa9fa18c79