ease the handling of the remote token url
when the user specifies the remote token url, he does not know, if
this url should terminate with a trailing slash or not. To prevent
the confusion, the last '/'s are stripped from the server url and
the left '/'s from the remote path are removed. Both parts are then
concatenated with a '/' in between.

Change-Id: I373e67fc6e237e4bcaf90a033b2700b9e25097c7