token: do not copy old pin on lost spass
During lost-token generation, the old pin is copied over.
However, in the case of spass token the pin is used as password,
and the user has lost it.

This commit avoids copying the old pin on lost spass events,
to make the new password alone usable when the previous spass-pin
is not anymore available.

In addition the manage ui disable the ability to enroll an
email or sms token as a lost token - only a pw token is allowed
where the password is set explicit

Change-Id: Iaa45f530150dfaeecd6abfeeece7950b1309eb4f
Signed-off-by: Luca Bruno <luca.bruno@rocket-internet.de>