token: do not copy old pin on lost spass
During lost-token generation, the old pin is copied over.
However, in the case of spass token the pin is used as password,
and the user has lost it.

This commit avoids copying the old pin on lost spass events,
to make the new password alone usable when the previous spass-pin
is not anymore available.

Signed-off-by: Luca Bruno <luca.bruno@rocket-internet.de>