linotpd: Fix for failing totptoken with empty timeshift value
due to the missing type check for values in the 2.7. releases
it was possible, that in the tokeninfo an empty value for the
timeshift was stored, which makes the totptoken unusable.
The problem was caused by the token default seetings, which
if interupted might insert Null as default value for the totp
default values.

closes #13640

Change-Id: If1e14ef402f66c6d2a4aff7a98687300d056d45d