linotpd: fix for the /admin/set which on error shows the escape content
As colateral of the escaping in javascript some dialogs are rendering
especialy exception with escape characters.
In case of errors in /admin/set the exception shows a message
which will be escaped in javascript and displayed in an alert
box. To beautify the content the quotes are replaced with different
chracters.

Change-Id: Ic88b0909f4616a66682b13833ece8bd5f165daa2