feature: support the migration of resolvers for assigned tokens
the migrate resolver tool supports the change of the uid inside the
resolver. This is done by adding a target resolver to the same
realm and specifying this as the target in the tools->migrate.
The migration works by searching all assigned tokens in the
src resolver and by looking for the login name of the user of
this token. Then the lookup of the login name in the target
resolver is made. If the user is found in the target resolver
his target resolver uid together with the target resolver name
is assigned to this token.

closes #13672

Change-Id: Ib86a1750629b2a53df7f8ab8cf627431359ed17e