reduce the amount of userid resolver password checks
with otppin=1 and a wrong password, LinOTP does too many resolver
password checks which could result in a final logouf of the user.
To prevent this, in case of otppin=1, the token validation share
an dict with the hashed passwords and the password check result.
One more optimization is made in the pass+opt validation, where
the otp verification will now be made first.

Change-Id: Ib5ea07274fe5aef8e3cc9c7e55b3675261495cef