failing decrpytion (mac compare) does not delete the backup data
a wrong password could not be detected other than by comparing
the mac of the encrypted token data. In case of this "data mismatch"
now a dedicated exception is raised, which results in not deleting
the backup data.
As well missing parameters are handled now to not delete the
backup data

Change-Id: I91a4217b2bb546d820e862c75f0aeb937ffe0c13