added check_status to support polling
for to support polling some bigger changes had to be made.
Before the challenge that was solved successfully was immediately
removed. By this change the challenges are removed only after
expiration.
In companion to this the challenge became a mac to prevent the abilty
to disable the second factor when having access to the database.
This mac was implemented by extending the HSM class to now support the
signing of messages.

Change-Id: Ic43a79bafbe4e33f29968559cadcb50575f43443