fix for admin/show with serial number and admin policy with wildcard
this fix handles two problems: one is that the admin policy with '*'
in action did not work, an additional 'show' was required. This was
caused by the additinal policy verification in the admin controller.
The other was, that in case of a given serial and no valid filter_realms,
the condition for the serial was removed instead of disabling the
query result by quering a token serial that does not exist, which
enforces an empty result.

closes #13930

Change-Id: I09ca6ed059ca53f6c07a148efe00a64842937da7