linotpd: debian: Remove debconf entry 'linotp/generate_enckey'
Replacing an existing encryption key (without backing up the previous key)
would lead to unreadable encrypted fields in the token database.

To prevent the possibility of replacing the encryption key on a LinOTP upgrade
by accident, the respective debconf question is removed.

During a fresh LinOTP installation a new key will be generated. However, if an
encryption key already exists, it will be left untouched.

closes #12721

Change-Id: I6cfd2d226d181c092d61266639aa2faa1c121a98