fix for breaking qrtoken tests
for the qrtoken functionality the ec keys are created on the first server
start. But during test runs the database / config entries are deleted
which as well deletes the ec keys. As the server is not restarted, the
ec keys are not recreated and the tests are failing.
The fix is to check for the ec keys now on every request.

Change-Id: I05d2067f5534216260db91347b651b259c39f9bb