fix for handling multiple users in selfservice authentication
It is possible to define a realm with multiple resolvers, which might
contain the same login user. For the administration and the authentication
this is no problem as the token is refering to the resolver. But in the
selfservice authentication the user is identified only by his password on
one resolver. The patch now ensures the following constrains:

* the login name is searched in all resolvers
* the password is verified in against all resolvers
* the userid (eg objectguid) must be the same in all resolvers

These constrain reflect the usecase, that the user might have different
roles, which are bundled in one realm.

closes #14195

Change-Id: Icce6e361bc4d12f0dff16d04e4e68adfdb3ce76d