smsprovider: fix encoding problem in SMPP provider
the smpp provider requires all input data to be of type
string (str) while in the context of th pylons request
all input is of type unicode. Problem is that the smpplib
treats str objects as byte string, which fails during
concatination when unicode is involved. In case of unicode
python tries to upgrade the byte string which will fail as
there is an incrementing sequence number involved. Thus all
input data is converted to iso8859-1 which is the smpp
data-coding=3. Other encoding is currently not supported!

closes #14206

Change-Id: Iefd4eb812eb8b42f1087b085983b0f6729f6f21e