fix the error with the tokencount handling during assignment
the tokencount policy check is done after the assignment of the
token, where the tokencount could be estimated correctly.
In this context the transaction commit to store the token
was moved to the past of the post policy check. In general
the to early commit operation in the model was replaced with
a flush operation.

closes #14304

Change-Id: I2139d34b55cdd7c6a011eacb0101e745b73375a6
(cherry picked from commit ce34518ce88a845e54486524d8dbf54344eaac41)