[LinOTP] redefined checkpins() and centralized checkpins call for manage-enroll
checkpins now expects one parameter that is either a jQuery selector string or an jQuery object.
 It is more flexibile this way to allow referencing more than two html ids
 but more than two elements as well as referencing classes or other selectors

a helper class "array_entries_equal" was extracted to linotp_utils to check if all
 entries of an array are "=="-compared equal

This allowed to remove all the listener-registrations in the templates for the most part.
They will now get assigned in tokentype_changed() where they are contentual reqiuired to be registered.

Change-Id: I26fe92b26168736ffa07351f1d04010378b2eaea