inital version of the new policy evaluation engine
the major design goal was to reduce the code replication.
Major reason for the large code base is the repeated filtering
as the old machine does the evaluation without state or agregation
af all filters.

beneath smaller and cleaner codebase it will be significantly faster :)
as the policies are added after evaluating all filters

this initial commit contains some migration of functions which repeatedly
called the getPolicy() as a prove for the statefull processing api which
will reduce the need for faulty post processing.

the code is restructured where the old processing has been moved to legacy.py
and the new code into processing.py. Functions, which are used by both modules
has been moved into util.py.

unit tests for the complex policy comparison functions has been added

Change-Id: I750cb22a2be2b8506ff00c8b7ea50ef26f8400db