Move the docker linotp builder sources to a volume
This allows building of linotp without needing to rebuild this
container every time. In addition, a developer can now build
the linotp image directly in their working directory.

The default build takes a copy of the mounted directory and builds
in there to avoid making changes to the working directory. This could
be removed later.

Change-Id: Iefe6e5d97f178dbfff1ad0e951979f13a17e29ba