fix for resolver test to return wrong number of users
in case of a copied resolver, the testconnection will recieve
no password and the lookup of the password returned a None
instead of an empty value, which failed during encoding. Thus
this fix simply set the return of the value lookup to an ''
empty string.

Beneath this a code rearangement was made, so that in case of
an noexisting local certfile wont fall back to the system certs.

closes #LINOTP-145

Change-Id: I9a0154e2326902c23f7bb84779e018c8539c87fd