fix for the ldap server uri handling
the ldap lib support an undocumented feature of uri containing multiple
server urls seperated by ',' and server urls with leading whitespaces.
But as we rely on differentiating the handling of ldaps:// and ldap://
we build code that did check the protocol prefix, which raised an
protocol error on leading whitespaces. The fix now splits the uri
for the testconnection in sepearte uris and runs test for each, finishing
on the first success. In the connect() method the uris are in addition
striped. The difference between the 2.9.0 is that now the connect()
method is used by both, the testconnection and the standard simple_bind,
due to the requirement to handle the certificates for ldaps cocrectly.

refers LinOTP-206

Change-Id: I5e7785651f45aefdb7323a2de45746c992e2a582